% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd.R
\name{accessor}
\alias{accessor}
\alias{data_dim}
\alias{ocdMethod}
\alias{n_obs}
\alias{patience}
\alias{param}
\alias{thresholds}
\alias{baselineMean}
\alias{baselineSD}
\alias{tracked}
\alias{statistics}
\alias{status}
\title{Accessor functions to attributes of class ChangepointDetector}
\usage{
data_dim(detector)

ocdMethod(detector)

n_obs(detector)

patience(detector)

param(detector)

thresholds(detector)

baselineMean(detector)

baselineSD(detector)

tracked(detector)

statistics(detector)

status(detector)
}
\arguments{
\item{detector}{Object of S3 class 'ChangepointDetector'}
}
\description{
Accessor functions to attributes of class ChangepointDetector
}
\details{
Obtain various attributes of the class 'ChangepointDetector'
}
\seealso{
\code{\link{ChangepointDetector}}
}
