% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd.R
\name{checkChange}
\alias{checkChange}
\title{Check if a mean change has occurred.}
\usage{
checkChange(detector)
}
\arguments{
\item{detector}{Object of class 'Changepoint Detector'}
}
\value{
Updated object \code{detector}
}
\description{
Check if a mean change has occurred.
}
\details{
The \code{\link{normalisedStatistics}} funcrtion is used to check if
any of the test statistics are above the threshold level. If this happens, the
status of the detector is changed to record the time of change and a message
is printed to the standard output declaring the change.
}
\seealso{
\code{\link{normalisedStatistics}}, \code{\link{setStatus}},
}
