% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{MC_Chan}
\alias{MC_Chan}
\title{Compute Monte Carlo thresholds for the Chan method}
\usage{
MC_Chan(dim, patience, p0, w, lambda, MC_reps)
}
\arguments{
\item{dim}{Data dimension}

\item{patience}{Nominal patience of the procedure}

\item{p0}{Assumed fraction of nonzero coordinates of change.}

\item{w}{Window size}

\item{lambda}{Tuning parameter for Chan (2017) mmethod}

\item{MC_reps}{number of Monte Carlo repetitions to use}
}
\value{
A numeric vector of computed thresholds.
}
\description{
Compute Monte Carlo thresholds for the Chan method
}
