% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{MC_Mei}
\alias{MC_Mei}
\title{Compute Monte Carlo thresholds for the Mei method}
\usage{
MC_Mei(dim, patience, b, MC_reps)
}
\arguments{
\item{dim}{Data dimension}

\item{patience}{Nominal patience of the procedure}

\item{b}{lLwer bound on per-coordinate magnitude of change}

\item{MC_reps}{Number of Monte Carlo repetitions to use}
}
\value{
A numeric vector of computed thresholds.
}
\description{
Compute Monte Carlo thresholds for the Mei method
}
