% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{MC_ocd}
\alias{MC_ocd}
\title{Compute Monte Carlo thresholds for the OCD method}
\usage{
MC_ocd(dim, patience, beta, sparsity, MC_reps)
}
\arguments{
\item{dim}{Data dimension}

\item{patience}{Nominal patience of the procedure}

\item{beta}{Lower bound on l_2 norm of change}

\item{sparsity}{Sparsity parameter for the OCD method}

\item{MC_reps}{Number of Monte Carlo repetitions to use}
}
\value{
A numeric vector of computed thresholds.
}
\description{
Compute Monte Carlo thresholds for the OCD method
}
