% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{MC_XS}
\alias{MC_XS}
\title{Compute Monte Carlo thresholds for the XS method}
\usage{
MC_XS(dim, patience, p0, w, MC_reps)
}
\arguments{
\item{dim}{Data dimension}

\item{patience}{Nominal patience of the procedure}

\item{p0}{Assumed fraction of nonzero coordinates of change.}

\item{MC_reps}{number of Monte Carlo repetitions to use}
}
\value{
A numeric vector of computed thresholds.
}
\description{
Compute Monte Carlo thresholds for the XS method
}
