% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{Mei_update}
\alias{Mei_update}
\title{Processing a new data point for the 'Mei' class}
\usage{
Mei_update(x_new, R, b)
}
\arguments{
\item{x_new}{a new data point}

\item{R}{vector of of tail CUSUMs that will be tracked and updated online}

\item{b}{a user specified lower bound on per-coordinate magnitude of the
vector of change to be detected.}
}
\value{
a list of
\itemize{
\item stat: a vector of 2 test statistics for the 'Mei' class.
\item R: the updated R vector
}
}
\description{
This function implements the \code{\link{getData}} function to
perform the online changepoint detection for the 'Mei' class.
}
