% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd.R
\name{new_Mei}
\alias{new_Mei}
\title{constructor of subclass 'Mei' in class 'ChangepointDetector'}
\usage{
new_Mei(dim, thresh, b)
}
\arguments{
\item{dim}{Data dimension, all new data must be of this dimension}

\item{thresh}{Detection threshold. A numeric vector of
length two (corresponding to the max and sum statistics) should be specified.}

\item{b}{Lower bound on the per-coordinate magnitude of mean change be
detected.}
}
\value{
An object of S3 subclass 'Mei' in class 'ChangepointDetector'.
}
\description{
constructor of subclass 'Mei' in class 'ChangepointDetector'
}
\details{
It is preferred to use \code{\link{ChangepointDetector}} for
construction.
}
\examples{
detector <- new_Mei(dim=100, thresh=c(8.6, 125.1), b=0.1)
}
