% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd.R
\name{normalisedStatistics}
\alias{normalisedStatistics}
\title{Compute maximum ratio between detection statistic and its threshold}
\usage{
normalisedStatistics(detector)
}
\arguments{
\item{detector}{Object of class 'Changepoint Detector'}
}
\value{
maximum of the ratio between the current test statistics and their
respective thresholds.
}
\description{
Compute maximum ratio between detection statistic and its threshold
}
