% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParkfieldSensor-data.R
\docType{data}
\name{ParkfieldSensor}
\alias{ParkfieldSensor}
\title{Parkfield seismic sensor data}
\format{A matrix with 39 columns and 14998 rows, with each column
corresponding to a sensor and each row corresponding to a time 
after 2am. Column names are seconds after 2am.}
\source{
HRSN (2014), High Resolution Seismic Network. 
UC Berkeley Seismological Laboratory. Dataset. doi:10.7932/HRSN.
}
\usage{
data(ParkfieldSensor)
}
\description{
Processed data from 39 ground motion sensors at 13 stations near 
Parkfield, California frm 2.00-2.16am on December 23, 2004, with
an earthquake measured at duration magnitude 1.47Md hit near
Atascadero, California at 02:09:54.01.
}
\keyword{datasets}
