% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd.R
\name{print.ChangepointDetector}
\alias{print.ChangepointDetector}
\title{Printing methods for the 'ChangepointDetector' class}
\usage{
\method{print}{ChangepointDetector}(x, ...)
}
\arguments{
\item{x}{object of the 'ChangepointDetector' class}

\item{...}{other arguments used in \code{print}}
}
\description{
Printing methods for the 'ChangepointDetector' class
}
