% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd.R
\name{reset}
\alias{reset}
\alias{reset.OCD}
\alias{reset.Mei}
\alias{reset.XS}
\alias{reset.Chan}
\title{Reset changepoint detector to initial state}
\usage{
reset(detector)

\method{reset}{OCD}(detector)

\method{reset}{Mei}(detector)

\method{reset}{XS}(detector)

\method{reset}{Chan}(detector)
}
\arguments{
\item{detector}{Object of class 'Changepoint Detector'}
}
\value{
Updated object \code{detector}
}
\description{
Reset changepoint detector to initial state
}
\section{Methods (by class)}{
\itemize{
\item \code{OCD}: Reset object of subclass 'OCD'

\item \code{Mei}: Reset object of subclass 'Mei'

\item \code{XS}: Reset object of subclass 'XS'

\item \code{Chan}: Reset object of subclass 'Chan'
}}

