% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd.R
\name{setBaselineMean}
\alias{setBaselineMean}
\title{Set baseline mean}
\usage{
setBaselineMean(detector, mean)
}
\arguments{
\item{detector}{Object of class 'Changepoint Detector'}

\item{mean}{vector of pre-change mean, must be of the same dimension as
specified in the data_dim attribute of \code{detector}.}
}
\value{
Updated object \code{detector}
}
\description{
Set baseline mean
}
