% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd.R
\name{setBaselineSD}
\alias{setBaselineSD}
\title{Set baseline standard deviation}
\usage{
setBaselineSD(detector, sd)
}
\arguments{
\item{detector}{Object of class 'Changepoint Detector'}

\item{sd}{vector of standard deviation, must be of the same dimension as
specified in the data_dim attribute of \code{detector}.}
}
\value{
Updated object \code{detector}
}
\description{
Set baseline standard deviation
}
