% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd.R
\name{setStatus}
\alias{setStatus}
\title{Set changepoint detector status}
\usage{
setStatus(detector, new_status)
}
\arguments{
\item{detector}{Object of class 'Changepoint Detector'}

\item{new_status}{'estimating' or 'monitoring'}
}
\value{
Updated object \code{detector}
}
\description{
Set changepoint detector status
}
\details{
If the status is set to 'estimating', new observations are used to
update current estimate of pre-change mean and standard deviation. If the 
status is set to 'monitoring', new observations are used to check if mean 
change has occurred.
}
