% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd.R
\name{update_param}
\alias{update_param}
\title{compute new mean and sd from old ones with one additional observation}
\usage{
update_param(old_mean, old_sd, x_new, n_obs)
}
\arguments{
\item{old_mean}{vector of old means}

\item{old_sd}{vector of old standard deviation}

\item{x_new}{new observation vector}

\item{n_obs}{total number of observations (including x_new)}
}
\value{
list of two vectors: new mean and new standard deviation
}
\description{
compute new mean and sd from old ones with one additional observation
}
